Fix: ltj_algorithm construction

+++ include/ltj_algorithm.hpp
@@ -242,11 +242,11 @@ namespace ring {
                 std::vector<ltj_iter_type*>& itrs = m_var_to_iterators[x_j];
                 for(ltj_iter_type* iter : itrs){
                     if(c == -1){
-                        std::cout << "Leap (min) with x_j=" << x_j << ": " << std::flush;
+                        std::cout << "Leap (min) with x_j=" << (uint64_t) x_j << ": " << std::flush;
                         c_i = iter->leap(x_j);
                         std::cout << c_i << std::endl;
                     }else{
-                        std::cout << "Leap (next) with x_j=" << x_j << ": " << std::flush;
+                        std::cout << "Leap (next) with x_j=" << (uint64_t) x_j << ": " << std::flush;
                         c_i = iter->leap(x_j, c);
                         std::cout << c_i << std::endl;
                     }
